Revert "Bot communication compiling. Needs testing"
