+++ Red_Bull_Main_Code/Red_Bull_Main_Code/Debug/Makefile
@@ -45,8 +45,7 @@ C_SRCS +=  \
 ../m_rf.c \
 ../m_usb.c \
 ../m_wii.c \
-../puck_location_function.c \
-../wireless_comms_function.c
+../puck_location_function.c
 
 
 PREPROCESSING_SRCS += 
@@ -64,8 +63,7 @@ m_bus.o \
 m_rf.o \
 m_usb.o \
 m_wii.o \
-puck_location_function.o \
-wireless_comms_function.o
+puck_location_function.o
 
 OBJS_AS_ARGS +=  \
 initialization_function.o \
@@ -76,8 +74,7 @@ m_bus.o \
 m_rf.o \
 m_usb.o \
 m_wii.o \
-puck_location_function.o \
-wireless_comms_function.o
+puck_location_function.o
 
 C_DEPS +=  \
 initialization_function.d \
@@ -88,8 +85,7 @@ m_bus.d \
 m_rf.d \
 m_usb.d \
 m_wii.d \
-puck_location_function.d \
-wireless_comms_function.d
+puck_location_function.d
 
 C_DEPS_AS_ARGS +=  \
 initialization_function.d \
@@ -100,8 +96,7 @@ m_bus.d \
 m_rf.d \
 m_usb.d \
 m_wii.d \
-puck_location_function.d \
-wireless_comms_function.d
+puck_location_function.d
 
 OUTPUT_FILE_PATH +=Red_Bull_Main_Code.elf
 
@@ -117,8 +112,6 @@ LINKER_SCRIPT_DEP+=
 
 
 # AVR32/GNU C Compiler
-
-
