+++ Red_Bull_Main_Code/Red_Bull_Main_Code/Debug/main.d
@@ -22,8 +22,7 @@ main.d main.o: .././main.c .././m_general.h \
  .././m_rf.h .././m_usb.h \
  c:\program\ files\ (x86)\atmel\studio\7.0\toolchain\avr8\avr8-gnu-toolchain\avr\include\stdlib.h \
  .././m_wii.h .././localization_function.h .././initialization_function.h \
- .././motor_control_function.h .././puck_location_function.h \
- .././wireless_comms_function.h
+ .././motor_control_function.h .././puck_location_function.h
 
 .././m_general.h:
 
@@ -82,5 +81,3 @@ c:\program\ files\ (x86)\atmel\studio\7.0\toolchain\avr8\avr8-gnu-toolchain\avr\
 .././motor_control_function.h:
 
 .././puck_location_function.h:
-
-.././wireless_comms_function.h:
