+++ Red_Bull_Main_Code/Red_Bull_Main_Code/Debug/makedep.mk
@@ -20,5 +20,3 @@ m_wii.c
 
 puck_location_function.c
 
-wireless_comms_function.c
-
