--- Red_Bull_Main_Code/Red_Bull_Main_Code/wireless_comms_function.h
@@ -1,11 +1,0 @@
-// Wireless Comms General Function
-
-#ifndef wireless_comms_function__
-#define wireless_comms_function__
-
-#include "m_general.h"
-#include "initialization_function.h"
-
-void wireless_send(int sending_bot, char* send_buffer); // Send communications
-
-#endif
